api/requirements.txt
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
aiofiles==24.1.0
cloudinary==1.41.0
